SpellEffects: Kompilierungswarnung gefixt.

+++ src/server/game/Spells/SpellEffects.cpp
@@ -3391,8 +3391,8 @@ void Spell::EffectDispel(SpellEffIndex effIndex)
 
     // If by any chance, there are less spells to dispel than the maximum number of dispellable spells, limit it.
     // IE Magic Dispel can dispel 2 spells, but we have only one dispellable buff, so limit the damage to 1.
-    if (damage > dispel_list.size())
-        damage = dispel_list.size();
+    if (damage > int32(dispel_list.size()))
+        damage = int32(dispel_list.size());
 
     // Ok if exist some buffs for dispel try dispel it
     uint32 failCount = 0;
